+++ src/dao/common.go
@@ -18,14 +18,17 @@ type Daos struct {
 	OperationLog *OperationLog
 	UrbsAcAcl    *UrbsAcAcl
 	UrbsAcUser   *UrbsAcUser
+	UrbsLock     *UrbsLock
 }
 
 // NewDaos ...
 func NewDaos(sql *service.SQL) *Daos {
-	daos := &Daos{
+	daos = &Daos{
 		OperationLog: &OperationLog{DB: sql.DB},
 		UrbsAcAcl:    &UrbsAcAcl{DB: sql.DB},
 		UrbsAcUser:   &UrbsAcUser{DB: sql.DB},
+
+		UrbsLock: &UrbsLock{DB: sql.DB},
 	}
 	return daos
 }