+++ CHANGELOG.md
@@ -10,3 +10,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 **Change:**
 
 - Add ac API.
+- Add test on github.
+- Add lock for sync group members.